Implement split and parallelize for get_datapoints_frame()

+++ cognite/timeseries.py
@@ -4,8 +4,8 @@
 This module mirrors the Timeseries API. It allows you to fetch data from the api and output it in various formats.
 """
 import io
-import warnings
 import os
+import warnings
 from datetime import datetime
 from functools import partial
 from multiprocessing import Pool
@@ -15,8 +15,8 @@ import pandas as pd
 import cognite._constants as _constants
 import cognite._utils as _utils
 import cognite.config as config
-from cognite.data_objects import DatapointsObject, LatestDatapointObject
 from cognite._protobuf_descriptors import _api_timeseries_data_v1_pb2
+from cognite.data_objects import DatapointsObject, LatestDatapointObject
 
 
 def get_datapoints(tag_id, aggregates=None, granularity=None, start=None, end=None, **kwargs):
@@ -53,17 +53,8 @@ def get_datapoints(tag_id, aggregates=None, granularity=None, start=None, end=No
         DatapointsObject: A data object containing the requested data with several getter methods with different
         output formats.
     '''
-
-
-
     api_key, project = config.get_config_variables(kwargs.get('api_key'), kwargs.get('project'))
-
-    if isinstance(start, datetime):
-        start = _utils.datetime_to_ms(start)
-    if isinstance(end, datetime):
-        end = _utils.datetime_to_ms(end)
-    if end is None:
-        end = _utils.get_last_datapoint_ts(tag_id, api_key=api_key, project=project)
+    start, end = _utils.interval_to_ms(start, end)
 
     diff = end - start
     num_of_processes = kwargs.get('processes', os.cpu_count())
@@ -86,14 +77,21 @@ def get_datapoints(tag_id, aggregates=None, granularity=None, start=None, end=No
 
     step_starts = [start + (i * step_size) for i in range(steps)]
     args = [{'start': start, 'end': start + step_size, 'display_progress': False} for start in step_starts]
+    display_progress = len(args) <= 2
+    if display_progress:
+        args[-1]['display_progress'] = True
+    else:
+        prog_ind = _utils.ProgressIndicator([tag_id], start, end, display_progress=False)
 
     p = Pool(steps)
 
     datapoints = p.map(partial_get_dps, args)
-
+    if not display_progress:
+        prog_ind.terminate()
     concat_dps = []
     [concat_dps.extend(el) for el in datapoints]
     return DatapointsObject({'data': {'items': [{'tagId': tag_id, 'datapoints': concat_dps}]}})
+
 
 def _get_datapoints_helper_wrapper(args, tag_id, aggregates, granularity, protobuf, api_key, project):
     return _get_datapoints_helper(
@@ -107,6 +105,7 @@ def _get_datapoints_helper_wrapper(args, tag_id, aggregates, granularity, protob
         project=project,
         display_progress=args['display_progress']
     )
+
 
 def _get_datapoints_helper(tag_id, aggregates=None, granularity=None, start=None, end=None, **kwargs):
     '''Returns a list of datapoints for the given query.
@@ -132,7 +131,7 @@ def _get_datapoints_helper(tag_id, aggregates=None, granularity=None, start=None
         protobuf (bool):        Download the data using the binary protobuf format. Only applicable when getting raw data.
                                 Defaults to True.
 
-        display_progress (bool):   Whether or not to display progress indicator.
+        display_progress (bool):   Whether or not to display progress indicator. Defaults to True.
 
         api_key (str):          Your api-key.
 
@@ -160,8 +159,8 @@ def _get_datapoints_helper(tag_id, aggregates=None, granularity=None, start=None
         'api-key': api_key,
         'accept': 'application/protobuf' if use_protobuf else 'application/json'
     }
-
-    prog_ind = _utils.ProgressIndicator([tag_id], start, end, api_key, project)
+    if kwargs.get('display_progress', True):
+        prog_ind = _utils.ProgressIndicator([tag_id], start, end, api_key, project)
     datapoints = []
     while not datapoints or len(datapoints[-1]) == limit:
         res = _utils.get_request(url, params=params, headers=headers)
@@ -173,17 +172,18 @@ def _get_datapoints_helper(tag_id, aggregates=None, granularity=None, start=None
             res = res.json()['data']['items'][0]['datapoints']
 
         if not res:
-            warning = "An interval with no data has been requested ({}, {})."\
+            warning = "An interval with no data has been requested ({}, {})." \
                 .format(params['start'], params['end'])
             warnings.warn(warning)
             break
 
         datapoints.append(res)
         latest_timestamp = int(datapoints[-1][-1]['timestamp'])
-        prog_ind.update_progress(latest_timestamp)
+        if kwargs.get('display_progress', True):
+            prog_ind.update_progress(latest_timestamp)
         params['start'] = latest_timestamp + (_utils.granularity_to_ms(granularity) if granularity else 1)
-
-    prog_ind.terminate()
+    if kwargs.get('display_progress', True):
+        prog_ind.terminate()
     dps = []
     [dps.extend(el) for el in datapoints]
     return dps
@@ -322,12 +322,104 @@ def get_datapoints_frame(tag_ids, aggregates, granularity, start=None, end=None,
                 ['<tagid1>', {'tagId': '<tag_id2>', 'aggregates': ['<aggfunc1>', '<aggfunc2>']}]
     '''
     api_key, project = config.get_config_variables(kwargs.get('api_key'), kwargs.get('project'))
-    url = config.get_base_url() + '/projects/{}/timeseries/dataframe'.format(project)
 
-    if isinstance(start, datetime):
-        start = _utils.datetime_to_ms(start)
-    if isinstance(end, datetime):
-        end = _utils.datetime_to_ms(end)
+    start, end = _utils.interval_to_ms(start, end)
+
+    diff = end - start
+    num_of_processes = kwargs.get('processes', os.cpu_count())
+
+    granularity_ms = 1
+    if granularity:
+        granularity_ms = _utils.granularity_to_ms(granularity)
+    steps = min(num_of_processes, max(1, int(diff / granularity_ms)))
+    step_size = _utils.round_to_nearest(int(diff / steps), base=granularity_ms)
+
+    partial_get_dpsf = partial(
+        _get_datapoints_frame_helper_wrapper,
+        tag_ids=tag_ids,
+        aggregates=aggregates,
+        granularity=granularity,
+        api_key=api_key,
+        project=project
+    )
+
+    step_starts = [start + (i * step_size) for i in range(steps)]
+    args = [{'start': start, 'end': start + step_size, 'display_progress': False} for start in step_starts]
+    display_progress = len(args) <= 2
+    if display_progress:
+        args[-1]['display_progress'] = True
+    else:
+        prog_ind = _utils.ProgressIndicator(tag_ids, start, end, display_progress=False)
+    p = Pool(steps)
+
+    dataframes = p.map(partial_get_dpsf, args)
+    if not display_progress:
+        prog_ind.terminate()
+    df = pd.concat(dataframes).drop_duplicates(subset='timestamp').reset_index(drop=True)
+    return df
+
+
+def _get_datapoints_frame_helper_wrapper(args, tag_ids, aggregates, granularity, api_key, project):
+    return _get_datapoints_frame_helper(
+        tag_ids,
+        aggregates,
+        granularity,
+        args['start'],
+        args['end'],
+        api_key=api_key,
+        project=project,
+        display_progress=args['display_progress']
+    )
+
+
+def _get_datapoints_frame_helper(tag_ids, aggregates, granularity, start=None, end=None, **kwargs):
+    '''Returns a pandas dataframe of datapoints for the given tag_ids all on the same timestamps.
+
+    This method will automate paging for the user and return all data for the given time period.
+
+    Args:
+        tag_ids (list):     The list of tag_ids to retrieve data for. Each tag_id can be either a string containing the
+                            tag_id or a dictionary containing the tag_id and a list of specific aggregate functions.
+
+        aggregates (list):  The list of aggregate functions you wish to apply to the data for which you have not
+                            specified an aggregate function. Valid aggregate functions are: 'average/avg, max, min,
+                            count, sum, interpolation/int, stepinterpolation/step'.
+
+        granularity (str):  The granularity of the aggregate values. Valid entries are : 'day/d, hour/h, minute/m,
+                            second/s', or a multiple of these indicated by a number as a prefix e.g. '12hour'.
+
+        start (Union[str, int, datetime]):    Get datapoints after this time. Format is N[timeunit]-ago where timeunit is w,d,h,m,s.
+                                    E.g. '2d-ago' will get everything that is up to 2 days old. Can also send time in ms since
+                                    epoch or a datetime object which will be converted to ms since epoch UTC.
+
+        end (Union[str, int, datetime]):      Get datapoints up to this time. Same format as for start.
+
+    Keyword Arguments:
+        api_key (str): Your api-key.
+
+        project (str): Project name.
+
+        display_progress (bool): Whether or not to display progress indicator.
+
+    Returns:
+        pandas.DataFrame: A pandas dataframe containing the datapoints for the given tag_ids. The datapoints for all the
+        tag_ids will all be on the same timestamps.
+
+    Note:
+        The ``tag_ids`` parameter can take a list of strings and/or dicts on the following formats::
+
+            Using strings:
+                ['<tag_id1>', '<tag_id2>']
+
+            Using dicts:
+                [{'tagId': '<tag_id1>', 'aggregates': ['<aggfunc1>', '<aggfunc2>']},
+                {'tagId': '<tag_id2>', 'aggregates': []}]
+
+            Using both:
+                ['<tagid1>', {'tagId': '<tag_id2>', 'aggregates': ['<aggfunc1>', '<aggfunc2>']}]
+    '''
+    api_key, project = kwargs.get('api_key'), kwargs.get('project')
+    url = config.get_base_url() + '/projects/{}/timeseries/dataframe'.format(project)
 
     num_aggregates = 0
     num_agg_per_tag = len(aggregates)
@@ -354,7 +446,8 @@ def get_datapoints_frame(tag_ids, aggregates, granularity, start=None, end=None,
         'accept': 'text/csv'
     }
     dataframes = []
-    prog_ind = _utils.ProgressIndicator(tag_ids, start, end, api_key, project)
+    prog_ind = _utils.ProgressIndicator(tag_ids, start, end, api_key, project,
+                                        display_progress=kwargs.get('display_progress', True))
     while not dataframes or dataframes[-1].shape[0] == per_tag_limit:
         res = _utils.post_request(url=url, body=body, headers=headers)
         dataframes.append(
